Fixes #61 Error when grunt.config has a circular dependency.
I encountered the issue webpack/grunt-webpack#61 today and set about
debugging. The plugin works fine when called directly from
`webpack-dev-server` on the command line, but throws an error when
called using grunt-karma. It appears the `grunt.config` function throws
an error if the the object has a nested circular dep.

Using lodash's `_.get` seems like the cleanest way to get this without
causing an error, and has a similar argument pattern to `grunt.config`.

Added a unit test for verification.