Fix for #90 - Error: Cannot find module
webpack actually uses relative paths so it is important to convert paths at least for loaders.

https://webpack.github.io/docs/configuration.html#module-loaders
```
IMPORTANT: The loaders here are resolved relative to the resource which they are applied to. This means they are not resolved relative the the configuration file. If you have loaders installed from npm and your node_modules folder is not in a parent folder of all source files, webpack cannot find the loader. You need to add the node_modules folder as absolute path to the resolveLoader.root option. (resolveLoader: { root: path.join(__dirname, "node_modules") })
```