feat: remove special handling of cache configuration
In webpack verion 4 special code was necessary to make the cache work with grunts multi-target system. The cache plugin in version 4 was storing one global cache for all compilers it was attached to. In webpack 5 this is not neccessary anymore as the cache now creates a new storage for each compiler.